[BUG] config: report unresolvable host names as errors
When a host name could not be resolved, an alert was emitted but the
service used to start with 0.0.0.0 for the IP address, because the
address parsing functions could not report an error. This is now
changed.

(cherry picked from commit d5191e77681b9c45f4ae36ea6eb051855d9e7675)