[MEDIUM] checks: add the ability to disable a server in the config
This is a backport of commit 9683909dce232b665b25017c6385c18021ab171f.
It makes it possible to disable a server by setting the "disabled"
keyword on its declaration line. The server will definitely be disabled
and no checks will be performed. It will be reported as "MAINT" in the
stats interface to indicate that this is not the result of a failed
check. This is also the term used in the equivalent feature in 1.4.