[BUG] client: don't add a new session to the list too early
During 1.3.16-dev, commit f54f8bdd8d2a40417e0fa03bdf8dfa5a4de73cfc
introduced a regression by creating a list of active sessions,
because a newly created session which is added to the list is
not removed if the request matches a monitor-net rule in TCP
mode. The result is that a dirty entry remains in the list,
later causing random crashes.

The fix consists in adding the session to the list only after
the monitor-net rules are checked. It must be backported to 1.3
and ported to 1.5-dev.

Thanks to Morten Gade Sørensen for reporting this bug with very
detailed diagnostics, and to Hervé Commowick for helping to find
a reproducer and isolating the affected versions. They also
confirmed the fix works.
(cherry picked from commit 71f84ded5197c94afac6bebe03cb4b37241e154e)