[BUG] config: disable "option forceclose" in non-http instances
Cyril Bouthors found that enabling "option forceclose" on a TCP mode
proxy caused the connection to be closed after the "timeout client"
strikes, even if there is activity, due to the way it works. This
problem is related to 1.3 only since forceclose has been changed in
1.4.

Forceclose is http-only but may be inherited from a defaults section,
so let's simply clear it during the validity checks if the proxy is
not in HTTP mode.